requirements.txt
django
djangorestframework
psycopg2-binary
requests
structlog
httpx
tenacity